+++ Navia.Engine/include/navia/platform/opengl/OpenGLRendererAPI.hpp
@@ -8,10 +8,10 @@ class OpenGLRendererAPI : public RendererAPI {
 public:
     void init() override;
     void shutdown() override;
-    void setViewport(size_t x, size_t y, size_t width, size_t height) override;
+    void setViewport(uint32_t x, uint32_t y, uint32_t width, uint32_t height) override;
     void setClearColor(const glm::vec4& color) override;
     void clear() override;
-    void drawIndexed(Ref<VertexArray> vertexArray, size_t indexCount) override;
+    void drawIndexed(Ref<VertexArray> vertexArray, uint32_t indexCount) override;
 };
 }
