+++ Navia.Engine/include/navia/renderer/RendererAPI.hpp
@@ -21,10 +21,10 @@ public:
 
     virtual void init() = 0;
     virtual void shutdown() = 0;
-    virtual void setViewport(size_t x, size_t y, size_t width, size_t height) = 0;
+    virtual void setViewport(uint32_t x, uint32_t y, uint32_t width, uint32_t height) = 0;
     virtual void setClearColor(const glm::vec4& color) = 0;
     virtual void clear() = 0;
-    virtual void drawIndexed(Ref<VertexArray> vertexArray, size_t indexCount) = 0;
+    virtual void drawIndexed(Ref<VertexArray> vertexArray, uint32_t indexCount) = 0;
 
 private:
     static APIVendor API;
