+++ Navia.Engine/include/navia/renderer/Shader.hpp
@@ -25,7 +25,7 @@ public:
     virtual void unbind() const = 0;
 
     virtual void setInt(const std::string& name, int value) = 0;
-    virtual void setIntArray(const std::string& name, int* array, size_t count) = 0;
+    virtual void setIntArray(const std::string& name, int* array, uint32_t count) = 0;
 
     virtual void setFloat(const std::string& name, float value) = 0;
     virtual void setFloat2(const std::string& name, const glm::vec2& vector) = 0;
