+++ Navia.Engine/include/navia/renderer/VertexBuffer.hpp
@@ -7,13 +7,13 @@
 namespace Navia {
 class VertexBuffer {
 public:
-    static Ref<VertexBuffer> create(size_t size);
-    static Ref<VertexBuffer> create(float* vertices, size_t size);
+    static Ref<VertexBuffer> create(uint32_t size);
+    static Ref<VertexBuffer> create(float* vertices, uint32_t size);
     virtual ~VertexBuffer() = default;
 
     virtual const BufferLayout& getLayout() const = 0;
     virtual void setLayout(const BufferLayout& layout) = 0;
-    virtual void setData(const void* data, size_t size) = 0;
+    virtual void setData(const void* data, uint32_t size) = 0;
 
     virtual void bind() const = 0;
     virtual void unbind() const = 0;
