+++ Navia.Engine/src/events/KeyEvent.cpp
@@ -1,15 +1,15 @@
 #include "navia/events/KeyEvent.hpp"
 
 namespace Navia {
-KeyEvent::KeyEvent(size_t keyCode) : keyCode(keyCode) {}
+KeyEvent::KeyEvent(uint32_t keyCode) : keyCode(keyCode) {}
 
-size_t KeyEvent::getKeyCode() const {
+uint32_t KeyEvent::getKeyCode() const {
     return keyCode;
 }
 
-KeyPressedEvent::KeyPressedEvent(size_t keyCode, size_t repeatCount) : KeyEvent(keyCode), repeatCount(repeatCount) {}
+KeyPressedEvent::KeyPressedEvent(uint32_t keyCode, uint32_t repeatCount) : KeyEvent(keyCode), repeatCount(repeatCount) {}
 
-size_t KeyPressedEvent::getRepeatCount() const {
+uint32_t KeyPressedEvent::getRepeatCount() const {
     return repeatCount;
 }
 
@@ -19,7 +19,7 @@ std::string KeyPressedEvent::toString() const {
     return stream.str();
 }
 
-KeyReleasedEvent::KeyReleasedEvent(size_t keyCode) : KeyEvent(keyCode) {}
+KeyReleasedEvent::KeyReleasedEvent(uint32_t keyCode) : KeyEvent(keyCode) {}
 
 std::string KeyReleasedEvent::toString() const {
     std::stringstream stream;
@@ -27,7 +27,7 @@ std::string KeyReleasedEvent::toString() const {
     return stream.str();
 }
 
-KeyTypedEvent::KeyTypedEvent(size_t keyCode) : KeyEvent(keyCode) {}
+KeyTypedEvent::KeyTypedEvent(uint32_t keyCode) : KeyEvent(keyCode) {}
 
 std::string KeyTypedEvent::toString() const {
     std::stringstream stream;
