+++ Navia.Engine/src/platform/opengl/OpenGLIndexBuffer.cpp
@@ -2,12 +2,12 @@
 #include <glad/glad.h>
 
 namespace Navia {
-OpenGLIndexBuffer::OpenGLIndexBuffer(size_t* indices, size_t count) : count(count) {
+OpenGLIndexBuffer::OpenGLIndexBuffer(uint32_t* indices, uint32_t count) : count(count) {
     NAVIA_PROFILE_FUNCTION();
 
     glCreateBuffers(1, &rendererId);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, rendererId);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(size_t), indices, GL_STATIC_DRAW);
+    glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(uint32_t), indices, GL_STATIC_DRAW);
 }
 
 OpenGLIndexBuffer::~OpenGLIndexBuffer() {
@@ -16,7 +16,7 @@ OpenGLIndexBuffer::~OpenGLIndexBuffer() {
     glDeleteBuffers(1, &rendererId);
 }
 
-size_t OpenGLIndexBuffer::getCount() const {
+uint32_t OpenGLIndexBuffer::getCount() const {
     return count;
 }
